+++ EngineGame/PrimitiveRenderer.h
@@ -21,7 +21,7 @@ public:
     void createLineIncremental(Point2D* p1, Point2D* p2);
     void createLine(Point2D* p1, Point2D* p2);
     void cretaeLineBySegment(LineSegment* s1);
-    void createCircle(Point2D* p1, int r);
+    void createCircle(Point2D* p1, float r);
     void createEllipse(Point2D* p1, float stAngle, float endAngle, int rx, int ry);
 
     void display(GLfloat vertices[], int GLShapeType, int verticesAmount);
